+++ src/test/java/com/epam/reportportal/junit/beforeafter/AfterClassFailedInParameterizedTest.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.epam.reportportal.junit.exception;
+package com.epam.reportportal.junit.beforeafter;
 
 import com.epam.reportportal.junit.ReportPortalListener;
-import com.epam.reportportal.junit.features.exception.ExpectedExceptionThrownTest;
+import com.epam.reportportal.junit.features.beforeafter.AfterClassFailedTwoParameterizedTest;
 import com.epam.reportportal.junit.utils.TestUtils;
 import com.epam.reportportal.listeners.ItemStatus;
 import com.epam.reportportal.service.ReportPortal;
@@ -27,42 +27,45 @@ import com.epam.ta.reportportal.ws.model.FinishTestItemRQ;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
-import org.mockito.ArgumentMatchers;
 
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.ArgumentMatchers.same;
 import static org.mockito.Mockito.*;
 
-public class ExpectedExceptionPassedTest {
+public class AfterClassFailedInParameterizedTest {
+	private static final int TEST_METHOD_NUMBER = 3;
 
 	private final String classId = CommonUtils.namedId("class_");
-	private final String methodId = CommonUtils.namedId("method_");
+	private final List<String> methodIds = Stream.generate(() -> CommonUtils.namedId("method_"))
+			.limit(TEST_METHOD_NUMBER)
+			.collect(Collectors.toList());
 
 	private final ReportPortalClient client = mock(ReportPortalClient.class);
 
 	@BeforeEach
 	public void setupMock() {
-		TestUtils.mockLaunch(client, null, null, classId, methodId);
+		TestUtils.mockLaunch(client, null, null, classId, methodIds);
 		TestUtils.mockBatchLogging(client);
 		ReportPortalListener.setReportPortal(ReportPortal.create(client, TestUtils.standardParameters()));
 	}
 
 	@Test
-	public void verify_correct_expected_exception_thrown() {
-		TestUtils.runClasses(ExpectedExceptionThrownTest.class);
+	public void verify_after_class_failure_in_parameterized_test() {
+		TestUtils.runClasses(AfterClassFailedTwoParameterizedTest.class);
 
-		verify(client, times(1)).startTestItem(ArgumentMatchers.startsWith("root_"), any());
-		verify(client, times(1)).startTestItem(same(classId), any());
-		ArgumentCaptor<FinishTestItemRQ> finishTestCaptor = ArgumentCaptor.forClass(FinishTestItemRQ.class);
-		verify(client, times(2)).finishTestItem(same(methodId), finishTestCaptor.capture());
-		ArgumentCaptor<FinishTestItemRQ> finishSuiteCaptor = ArgumentCaptor.forClass(FinishTestItemRQ.class);
-		verify(client, times(1)).finishTestItem(same(classId), finishSuiteCaptor.capture());
+		verify(client, times(TEST_METHOD_NUMBER)).startTestItem(same(classId), any());
 
-		List<FinishTestItemRQ> items = finishTestCaptor.getAllValues();
-		assertThat(items.get(0).getStatus(), equalTo(ItemStatus.FAILED.name()));
-		assertThat(items.get(1).getStatus(), equalTo(ItemStatus.PASSED.name()));
+		ArgumentCaptor<FinishTestItemRQ> finishCaptor = ArgumentCaptor.forClass(FinishTestItemRQ.class);
+		methodIds.forEach(id -> verify(client, times(1)).finishTestItem(same(id), finishCaptor.capture()));
+
+		List<FinishTestItemRQ> finishRqs = finishCaptor.getAllValues();
+
+		finishRqs.subList(0, 2).forEach(rq -> assertThat(rq.getStatus(), equalTo(ItemStatus.PASSED.name())));
+		assertThat(finishRqs.get(2).getStatus(), equalTo(ItemStatus.FAILED.name()));
 	}
 }